+++ R/FINEMAP_check_existing_results.R
@@ -1,19 +1,25 @@
-FINEMAP_check_existing_results <- function(locus_dir,
+FINEMAP_check_existing_results <- function(dat,
+                                           locus_dir,
                                            credset_thresh,
                                            finemap_version,
                                            master_path,
                                            force_new = FALSE,
                                            verbose = TRUE){
-    if(any(file.exists(file.path(dirname(master_path),
-                                 c("data.cred","data.snp","data.config")))) &&
-       force_new==FALSE){
+    
+    file_options <- FINEMAP_check_files(locus_dir = locus_dir) 
+    if((length(file_options)>0) &&
+       isFALSE(force_new)){
         messager("Importing pre-computed FINEMAP results files.",
                  "Set force_new=TRUE to compute new results.",v=verbose)
-        dat <- FINEMAP_process_results(locus_dir = locus_dir,
-                                       dat = dat,
-                                       credset_thresh = credset_thresh,
-                                       results_file = ".cred",
-                                       finemap_version = finemap_version)
+        ## Use tryCatch because sometimes the requested version of FINEMAP
+        ## if different from the version that produced the results.
+        dat <- tryCatch({
+            FINEMAP_process_results(dat = dat,
+                                    locus_dir = locus_dir,
+                                    credset_thresh = credset_thresh, 
+                                    finemap_version = finemap_version,
+                                    verbose = verbose)
+        }, error=function(e){message(e);NULL})
         return(dat)
     } else {return(NULL)}
 }