+++ R/FINEMAP_find_executable.R
@@ -17,38 +17,20 @@ FINEMAP_find_executable <- function(FINEMAP_path=NULL,
                                             which = "cache"),
                                         "FINEMAP"
                                     ),
-                                    version="1.4",
+                                    version=package_version("1.4.1"),
                                     verbose=TRUE){
-    if(is.null(OS)){OS <- get_os()}
-    
-    if(version=="1.4"){
-        messager("+ Using FINEMAP v1.4",v=verbose)
-        if(OS=="osx"){
-            exec <- "finemap_v1.4_MacOSX"
-        } else{
-            exec <- "finemap_v1.4_x86_64"
-        }
-    }
-    if(version=="1.3.1") {
-        messager("+ Using FINEMAP v1.3.1",v=verbose)
-        if(OS=="osx"){
-            exec <- "finemap_v1.3.1_MacOSX"
-        } else{
-            exec <- "finemap_v1.3.1_x86_64"
-        }
-    }
-    if(version=="1.3") {
-        messager("+ Using FINEMAP v1.3",v=verbose)
-        if(OS=="osx"){
-            exec <- "finemap_v1.3_MacOSX"
-        } else{
-            exec <- "finemap_v1.3_x86_64"
-        }
-    }
+    if(is.null(OS)) OS <- get_os()
+    # messager("+ Using FINEMAP",paste0("v",version),v=verbose)
+    if(OS=="osx"){
+        exec <- paste0("finemap_v",version,"_MacOSX")
+    } else{
+        exec <- paste0("finemap_v",version,"_x86_64")
+    }   
     #### (Download and) test executable ###
-    if(is.null(FINEMAP_path)) FINEMAP_path <- file.path(save_dir,exec,exec) 
-    if(!file.exists(FINEMAP_path)){
-        dir.create(save_dir, showWarnings = FALSE, recursive = TRUE)
+    if(is.null(FINEMAP_path)) {
+        FINEMAP_path <- file.path(save_dir,exec,exec) 
+    }
+    if(!file.exists(FINEMAP_path)){ 
         tgz <- get_data(fname = paste0(exec,".tgz"), 
                         storage_dir = save_dir) 
         utils::untar(tarfile = tgz,