+++ R/FINEMAP_process_results.R
@@ -15,64 +15,74 @@
 FINEMAP_process_results <- function(dat,
                                     locus_dir,
                                     credset_thresh=.95,
-                                    pvalue_thresh=.05,
-                                    n_causal=5,
-                                    finemap_version=package_version("1.4"),
-                                    results_file=".cred",
-                                    nThread=1,
+                                    pvalue_thresh=.05, 
+                                    finemap_version=package_version("1.4.1"),
+                                    file_options=NULL,
                                     sort_by_CS=TRUE,
                                     verbose=TRUE){
-    PP <- NULL;
+    PP <- rsid <- prob <- NULL;
     #### Notes on FINEMAP output files ####
     ##
     ## .snp and .cred are often similar, but not identical.
     ## Recommendation: use the .cred file that shows the largest posterior 
     ## probability for the number of causal variants in line 1 of the file.
     ## and extract credible sets from that file. 
+    pp_cols <- c("PP","PP_snp","PP_config","CS","k") 
+    pp_cols <- pp_cols[pp_cols %in% names(dat)]
+    if(length(pp_cols)>0) dat <- dat[,-pp_cols, with=FALSE]
     
-    #### Handling FINEMAP version differences  ####
-    results_file <- FINEMAP_check_version_results(
-        finemap_version = finemap_version,
-        results_file = results_file)
     #### Double check which results files are available ####
     ## This vary depending on which version of FINEMAP you're using.
-    results_file <- FINEMAP_check_files(locus_dir, results_file)
-    #### Process FINEMAP results ####
-    if(results_file==".cred"){
+    file_options <- FINEMAP_check_files(locus_dir = locus_dir, 
+                                        file_options = file_options) 
+    #### Add conditional probabilities ####
+    if(".cred" %in% file_options){
         data.cred <- FINEMAP_import_data_cred(
             locus_dir = locus_dir, 
+            credset_thresh = credset_thresh, 
+            pvalue_thresh = pvalue_thresh,
             finemap_version = finemap_version,
             verbose = verbose) 
-        dat2 <- data.table::merge.data.table(data.table::data.table(dat),
-                                            data.table::data.table(data.cred),
-                                            by="SNP",
-                                            all.x = TRUE)
-    } else if (results_file==".snp"){
+        dat <- data.table::merge.data.table(
+            data.table::data.table(dat),
+            data.table::data.table(data.cred),
+            by="SNP",
+            all.x = TRUE)
+    } 
+    #### Add marginal probabilities ####
+    if (".snp" %in% file_options){
         data.snp <- FINEMAP_import_data_snp(
+            credset_thresh = credset_thresh,
             locus_dir = locus_dir,
             verbose = verbose) 
-        dat2 <- data.table::merge.data.table(data.table::data.table(dat),
-                                            data.table::data.table(
-                                                subset(data.snp, 
-                                                       select=c("rsid",
-                                                                "PP",
-                                                                "CS")) ),
-                                            by.x = "SNP",
-                                            by.y="rsid",
-                                            all.x = TRUE)
-    } else if (results_file==".config"){
+        dat <- data.table::merge.data.table(
+            data.table::data.table(dat),
+            data.table::data.table(
+                dplyr::select(data.snp,
+                              SNP=rsid,
+                              PP.snp=prob)),
+            by = "SNP", 
+            all.x = TRUE)
+    } 
+    #### Add configuration probabilities ####
+    if (".config" %in% file_options){
         data.config <- FINEMAP_import_data_config(
             locus_dir = locus_dir,
+            credset_thresh = credset_thresh,
+            pvalue_thresh = pvalue_thresh,
             finemap_version = finemap_version,
             verbose = verbose)
-        dat2 <- data.table::merge.data.table(data.table::data.table(dat),
-                                            data.table::data.table(data.config),
-                                            by="SNP",
-                                            all.x = TRUE)
+        dat <- data.table::merge.data.table(
+            data.table::data.table(dat),
+            data.table::data.table(data.config) %>%
+                dplyr::rename(PP.config=prob),
+            by="SNP",
+            all.x = TRUE)
     }
     #### Sort so that CS SNPs are at the top ####
-    if(sort_by_CS){
-        dat2 <- dat2 %>% dplyr::arrange(dplyr::desc(PP))
+    if(sort_by_CS){ 
+        dat <- dplyr::arrange(dat, dplyr::desc(PP))
     }
-    return(dat2)
+    return(dat)
 }
+